remove ErgoAddress and use the one from sigma; move json encoder/decoder into ErgoAddressJsonEncoder;
