+++ src/main/scala/org/ergoplatform/nodeView/wallet/ErgoWalletActor.scala
@@ -45,7 +45,7 @@ class ErgoWalletActor(settings: ErgoSettings) extends Actor with ScorexLogging {
   private var height = 0
   private var lastBlockUtxoRootHash = ADDigest @@ Array.fill(32)(0: Byte)
 
-  private implicit val addressEncoder = ErgoAddressEncoder(settings)
+  private implicit val addressEncoder = ErgoAddressEncoder(settings.chainSettings.addressPrefix)
   private val publicKeys: Seq[P2PKAddress] = Seq(prover.dlogPubkeys: _ *).map(P2PKAddress.apply)
 
   private val trackedAddresses: mutable.Buffer[ErgoAddress] = publicKeys.toBuffer
