+++ src/main/scala/org/ergoplatform/nodeView/wallet/ErgoWalletReader.scala
@@ -5,6 +5,7 @@ import java.util.concurrent.TimeUnit
 import akka.actor.ActorRef
 import akka.pattern.ask
 import akka.util.Timeout
+import org.ergoplatform.{ErgoAddress, P2PKAddress}
 import org.ergoplatform.modifiers.mempool.ErgoTransaction
 import org.ergoplatform.nodeView.wallet.ErgoWalletActor.{GenerateTransaction, ReadPublicKeys, ReadRandomPublicKey, ReadTrackedAddresses}
 import scorex.core.transaction.wallet.VaultReader
