+++ src/test/scala/org/ergoplatform/utils/WalletGenerators.scala
@@ -1,5 +1,6 @@
 package org.ergoplatform.utils
 
+import org.ergoplatform.{ErgoAddressEncoder, Pay2SAddress}
 import org.ergoplatform.modifiers.mempool.ErgoTransaction
 import org.ergoplatform.nodeView.wallet._
 import org.ergoplatform.settings.ErgoSettings
@@ -9,7 +10,7 @@ import sigmastate.Values
 trait WalletGenerators extends ErgoTransactionGenerators {
 
   private val ergoSettings = ErgoSettings.read(None)
-  private implicit val ergoAddressEncoder = new ErgoAddressEncoder(ergoSettings)
+  private implicit val ergoAddressEncoder = ErgoAddressEncoder(ergoSettings.chainSettings.addressPrefix)
 
 
   def trackedBoxGen: Gen[TrackedBox] = {
